requirements.txt
Pillow==9.4.0
PyYAML==6.0
Requests==2.32.3
